requirements.txt
amqp==5.2.0
Appium-Python-Client==4.1.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cycler==0.12.1
Django==5.1
django-rosetta==0.10.0
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
fonttools==4.53.1
h11==0.14.0
idna==3.7
kiwisolver==1.4.5
kombu==5.4.0
mysqlclient==2.2.4
numpy==2.1.0
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
polib==1.2.0
prompt_toolkit==3.0.47
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
